Update examples/braf-v600.yaml
Good to know, thank you! I'll take it that we should _only_ include the expressions for the subject Allele itself. Larger question - is it the intent / kosher for related Alleles that translate to an amino acid that satisfy the subject catvar be included under the `members` property? e.g., hgvs.g Alleles that translate to a BRAF V600 amino acid change.

Co-authored-by: Alex H. Wagner, PhD <Alex.Wagner@nationwidechildrens.org>